+++ index.html
@@ -4,6 +4,8 @@
     </head>
     <body>
         <main id="app"></main>
+        <link rel="stylessheet" href="/src/style/reset.css" >
+        <link rel="stylessheet" href="/src/style/style.css" >
         <script src="/src/main.js" type = "module"></script>
     </body>
 </html>
