+++ src/EditPage/editor.js
@@ -21,7 +21,7 @@ export default function Editor({
         console.log(nextState)
         $editor.querySelector('[name=title]').value = this.state.title
         $editor.querySelector('[name=content]').value = this.state.content
-        //this.render()
+        this.render()
     }
 
     this.render = () => {
