+++ src/linkButton.js
@@ -1,4 +1,4 @@
-import {push} from "./router.js"
+import {push} from "./utils/router.js"
 
 
 export default function LinkButton({
