+++ src/utils/storage.js
@@ -3,6 +3,7 @@ const storage =  window.localStorage
 export const getItem = (key,defaultValue) => {
     try {
         const storageValue = storage.getItem(key)
+        console.log("확인",storageValue ? JSON.parse(storageValue) : defaultValue)
         return storageValue ? JSON.parse(storageValue) : defaultValue
     } catch(e) {
         return defaultValue
